requirements.txt
django==5.0.1
python-dotenv==1.0.0
google-cloud-speech==2.24.1
google-cloud-storage==2.14.0
